Assign a total score to each book and return the most relevant hits (#17)
Assign a total score to each book as a weighted sum and return the most relevant (ranked) search results based on query and user info.

The total score consists of a title score with an assigned weight, a genre score and a description score (all three based on query) as well as a user preference bonus that is query independent.

The user preference scoring in BookRecommender.java is based on title and genre overlap with liked books

Note: Book ID matching with Elasticsearch is not yet functional yet and requires fix

Made title2id public in Database.java to allow external access for ID lookups

Updated intid to Integer from int in User.java to support nullability and interoperability

Add a subset of random books both query and user independent to the search results

These changes are part of ongoing work to retrieve search hits that feel relevant to the query as well as personalize search results based on user's liked books, with some random surfing to make sure to return recommendations even if there are few or no relevant matches.

---------

Co-authored-by: calvileong <clvleong@gmail.com>
Co-authored-by: Kajsa Lidin <Kajlid@users.noreply.github.com>